Replace old SLAM in LidarView superbuild by this one
See merge request keu-computervision/slam!22

Replace the old LiDAR SLAM currently in LidarView by this one.

Major changes :

* The SLAM lib `LidarSlam` and paraview plugin `LidarSlamPlugin` can now be correctly installed (binaries, libs and header files)
* CMake idefault build type and C++ standard
* Modifications on dependencies finding and linking
* Compilation warnings corrections

Related merge requests (please merge them in this following order) :

* [LiDAR SLAM](https://gitlab.kitware.com/keu-computervision/slam/-/merge_requests/22) (this MR)
* [LidarView-Superbuild](https://gitlab.kitware.com/LidarView/lidarview-superbuild/-/merge_requests/6)
* [LidarView-Core](https://gitlab.kitware.com/LidarView/lidarview-core/-/merge_requests/57)
* [LidarView](https://gitlab.kitware.com/LidarView/lidarview/-/merge_requests/25)

+++ slam_lib/include/LidarSlam/SpinningSensorKeypointExtractor.h
@@ -20,7 +20,7 @@
 #ifndef SpinningSensorKeypointExtractor_H
 #define SpinningSensorKeypointExtractor_H
 
-#include "LidarPoint.h"
+#include "LidarSlam/LidarPoint.h"
 
 #include <pcl/point_cloud.h>
 
@@ -120,7 +120,7 @@ private:
   void SetKeyPointsLabels();
 
   // Check if scanLine is almost empty
-  inline bool IsScanLineAlmostEmpty(size_t nScanLinePts) const { return nScanLinePts < 2 * this->NeighborWidth + 1; }
+  inline bool IsScanLineAlmostEmpty(int nScanLinePts) const { return nScanLinePts < 2 * this->NeighborWidth + 1; }
 
   // ---------------------------------------------------------------------------
   //   Parameters
